Update the index and adding HTML, Digital Electronic courses in free-programming-books-hi.md
Sir I updated the index by adding the C context in index and also i added the notes of HTML and Digital electronic in Hindi. Sir Digital Electronic is a important course that are covered by students in Engineering and I genuinely wanted to be in this project and sir HTML is a basic course that every student search for, so HTML in Hindi might help those students who seek for help in html. Thank you sir. Hope so my changes will might help. :)